// src/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
};
// ...
export enum BonusLevelType {
  Junior = 'JUNIOR',
  Middle = 'MIDDLE',
  Senior = 'SENIOR'
}

export type BonusTicket = {
  __typename?: 'BonusTicket';
  activeTill?: Maybe<Scalars['DateTime']['output']>;
  code: Scalars['String']['output'];
  createdAt: Scalars['DateTime']['output'];
  id: Scalars['String']['output'];
  isActive: Scalars['Boolean']['output'];
  ticketType: BonusTicketType;
  updatedAt: Scalars['DateTime']['output'];
};

export enum BonusTicketType {
  Const = 'CONST',
  Disposable = 'DISPOSABLE'
}
// ...
export enum BonusValueType {
  Percent = 'PERCENT',
  Point = 'POINT'
}
// ...
export type MutationActivateBonusArgs = {
  bonusTicketId: Scalars['String']['input'];
  userId: Scalars['String']['input'];
};


export type MutationAuthGoogleArgs = {
  token: Scalars['String']['input'];
};


export type MutationCreateBonusArgs = {
  asset: Scalars['String']['input'];
  condition?: InputMaybe<Scalars['Float']['input']>;
  count: Scalars['Float']['input'];
  description: Scalars['String']['input'];
  isActive?: InputMaybe<Scalars['Boolean']['input']>;
  level?: InputMaybe<BonusLevelType>;
  payload?: InputMaybe<Scalars['JSON']['input']>;
  valueType: BonusValueType;
};


export type MutationCreateBonusTicketArgs = {
  activeTill?: InputMaybe<Scalars['DateTime']['input']>;
  bonusId: Scalars['String']['input'];
  code: Scalars['String']['input'];
  ticketType?: InputMaybe<BonusTicketType>;
  userId: Scalars['String']['input'];
};
// ...
export type MutationDeleteBonusArgs = {
  id: Scalars['String']['input'];
};


export type MutationDeleteBonusTicketArgs = {
  id: Scalars['String']['input'];
};
// ...
export type MutationUpdateBonusArgs = {
  asset?: InputMaybe<Scalars['String']['input']>;
  condition?: InputMaybe<Scalars['Float']['input']>;
  count?: InputMaybe<Scalars['Float']['input']>;
  description?: InputMaybe<Scalars['String']['input']>;
  id: Scalars['String']['input'];
  isActive?: InputMaybe<Scalars['Boolean']['input']>;
  level?: InputMaybe<BonusLevelType>;
  payload?: InputMaybe<Scalars['JSON']['input']>;
  valueType?: InputMaybe<BonusValueType>;
};


export type MutationUpdateBonusTicketArgs = {
  activeTill?: InputMaybe<Scalars['DateTime']['input']>;
  bonusId?: InputMaybe<Scalars['String']['input']>;
  code?: InputMaybe<Scalars['String']['input']>;
  id: Scalars['String']['input'];
  ticketType?: InputMaybe<BonusTicketType>;
  userId?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type QueryFindBonusArgs = {
  asset?: InputMaybe<Scalars['String']['input']>;
  id?: InputMaybe<Scalars['String']['input']>;
  isActive?: InputMaybe<Scalars['Boolean']['input']>;
  level?: InputMaybe<BonusLevelType>;
  skip?: InputMaybe<Scalars['Float']['input']>;
  take?: InputMaybe<Scalars['Float']['input']>;
  valueType?: InputMaybe<BonusValueType>;
};
// ...
export type QueryGetTicketArgs = {
  code: Scalars['String']['input'];
};